requirements.txt
Flask>=2.3.0
python-dotenv>=1.0.0
strands-agents
strands-agents-tools
strands-agents-builder
boto3>=1.28.0
psycopg2-binary>=2.9.0
